# deps: updates TinyGo to v0.28.1